image_backend/requirements.txt
facenet_pytorch==2.5.3
Flask==3.0.0
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.0
Flask_JWT_Extended==4.5.3
Flask_Mail==0.9.1
minio==7.1.17
mongoengine==0.27.0
numpy==1.24.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.74
pandas==1.4.2
Pillow==10.1.0
Pillow==10.1.0
pyunpack==0.3
Requests==2.31.0
torch==1.12.1
torchvision==0.13.1
yolov5==7.0.13
yolov8==0.0.2
